+++ test/test_source_base.py
@@ -23,7 +23,7 @@ from .datastore import datastore, registry
 # a download order.
 SourceAndFiles = {
     'bioformats': {
-        'read': r'\.(czi|jp2|svs|scn|dcm|qptiff)$',
+        'read': r'(\.(czi|jp2|svs|scn|dcm|qptiff)|[0-9a-f].*\.dcm)$',
         'noread': r'JK-kidney_B',
         'skip': r'TCGA-AA-A02O.*\.svs',
         # We need to modify the bioformats reader similar to tiff's
@@ -36,13 +36,13 @@ SourceAndFiles = {
     },
     'dummy': {'any': True, 'skipTiles': r''},
     'gdal': {
-        'read': r'\.(jpg|jpeg|jp2|ptif|scn|svs|tif.*|qptiff)$',
+        'read': r'(\.(jpg|jpeg|jp2|ptif|scn|svs|tif.*|qptiff)|18[-0-9a-f]{34}\.dcm)$',
         'noread': r'(huron\.image2_jpeg2k|sample_jp2k_33003|TCGA-DU-6399|\.(ome.tiff|nc)$)',
         'skip': r'nokeyframe\.ome\.tiff$',
         'skipTiles': r'\.*nc$',
     },
     'mapnik': {
-        'read': r'\.(jpg|jpeg|jp2|ptif|nc|scn|svs|tif.*|qptiff)$',
+        'read': r'(\.(jpg|jpeg|jp2|ptif|nc|scn|svs|tif.*|qptiff)|18[-0-9a-f]{34}\.dcm)$',
         'noread': r'(huron\.image2_jpeg2k|sample_jp2k_33003|TCGA-DU-6399|\.(ome.tiff)$)',
         'skip': r'nokeyframe\.ome\.tiff$',
         # we should only test this with a projection
@@ -67,24 +67,24 @@ SourceAndFiles = {
         'skipTiles': r'one_layer_missing',
     },
     'pil': {
-        'read': r'\.(jpg|jpeg|png|tif.*)$',
-        'noread': r'(G10-3|JK-kidney|d042-353.*tif|huron|one_layer_missing|US_Geo|extraoverview)',
+        'read': r'(\.(jpg|jpeg|png|tif.*)|18[-0-9a-f]{34}\.dcm)$',
+        'noread': r'(G10-3|JK-kidney|d042-353.*tif|huron|one_layer_missing|US_Geo|extraoverview)',  # noqa
     },
     'rasterio': {
-        'read': r'\.(jpg|jpeg|jp2|ptif|scn|svs|tif.*|qptiff)$',
+        'read': r'(\.(jpg|jpeg|jp2|ptif|scn|svs|tif.*|qptiff)|18[-0-9a-f]{34}\.dcm)$',
         'noread': r'(huron\.image2_jpeg2k|sample_jp2k_33003|TCGA-DU-6399|\.(ome.tiff|nc)$)',
         'skip': r'nokeyframe\.ome\.tiff$',
     },
     'test': {'any': True, 'skipTiles': r''},
     'tiff': {
-        'read': r'\.(ptif|scn|svs|tif.*|qptiff)$',
+        'read': r'(\.(ptif|scn|svs|tif.*|qptiff)|[-0-9a-f]{36}\.dcm)$',
         'noread': r'(oahu|DDX58_AXL|G10-3_pelvis_crop|'
                   r'd042-353\.crop\.small\.float|landcover_sample|US_Geo\.|'
                   r'imagej|bad_axes|nokeyframe\.ome\.tiff$)',
         'skipTiles': r'(sample_image\.ptif|one_layer_missing_tiles)'},
     'tifffile': {
         'read': r'',
-        'noread': r'((\.(nc|nd2|yml|yaml|json|czi|png|jpg|jpeg|jp2|dcm|zarr\.db|zarr\.zip)|(nokeyframe\.ome\.tiff|XY01\.ome\.tif)$)' +  # noqa
+        'noread': r'((\.(nc|nd2|yml|yaml|json|czi|png|jpg|jpeg|jp2|zarr\.db|zarr\.zip)|(nokeyframe\.ome\.tiff|XY01\.ome\.tif|level.*\.dcm)$)' +  # noqa
                   (r'|bad_axes' if sys.version_info < (3, 9) else '') +
                   r')',
     },